Implemented new feature

+++ src/main/java/com/example/movie_management/movie/Movie.java
@@ -9,7 +9,7 @@ public class Movie{
     public boolean adult;
     public String backdrop_path;
     public BelongsToCollection belongs_to_collection;
-    public int budget;
+    public long budget;
     public List<Genre> genres;
     public String homepage;
     public int id;
@@ -22,7 +22,7 @@ public class Movie{
     public List<ProductionCompany> production_companies;
     public List<ProductionCountry> production_countries;
     public String release_date;
-    public int revenue;
+    public long revenue;
     public int runtime;
     public List<SpokenLanguage> spoken_languages;
     public String status;
